fastapi/requirements.txt
fastapi
pydantic
uvicorn
requests
cassandra-driver